+++ packages/components/autocomplete/stories/Autocomplete.stories.tsx
@@ -4,6 +4,7 @@ import type { Meta } from '@storybook/react/types-6-0';
 import { Stack } from '@contentful/f36-core';
 import { FormControl } from '@contentful/f36-forms';
 import { Paragraph } from '@contentful/f36-typography';
+import { Button } from '@contentful/f36-button';
 import { getStringMatch } from '@contentful/f36-utils';
 
 import { Autocomplete } from '../src/Autocomplete';
@@ -81,7 +82,7 @@ export const Basic = (args: AutocompleteProps<string>) => {
 
   return (
     <Stack
-      style={{ minWidth: '300px' }}
+      style={{ width: '150px' }}
       flexDirection="column"
       spacing="spacingM"
       alignItems="start"
@@ -104,7 +105,10 @@ Basic.args = {
 };
 
 export const UsingObjectsAsItems = (args: AutocompleteProps<Produce>) => {
-  const [selectedFruit, setSelectedFruit] = useState<Produce>();
+  const [selectedFruit, setSelectedFruit] = useState<Produce>({
+    id: undefined,
+    name: '',
+  });
   const [filteredItems, setFilteredItems] = useState(fruits);
 
   const handleInputValueChange = (value: string) => {
@@ -143,8 +147,54 @@ UsingObjectsAsItems.args = {
   placeholder: 'Search your favorite fruit',
 };
 
-export const UsingGroupedItems = (args: AutocompleteProps<GroceryList>) => {
-  const [selectedItem, setSelectedItem] = useState<Produce>();
+export const ControlledFromOutside = (args: AutocompleteProps<Produce>) => {
+  const [selectedFruit, setSelectedFruit] = useState<Produce>({
+    id: 9,
+    name: 'Pear 🍐',
+  });
+  const [filteredItems, setFilteredItems] = useState(fruits);
+
+  const handleInputValueChange = (value: string) => {
+    const newFilteredItems = fruits.filter((item) =>
+      item.name.toLowerCase().includes(value.toLowerCase()),
+    );
+    setFilteredItems(newFilteredItems);
+  };
+
+  const handleSelectItem = (item: Produce) => {
+    setSelectedFruit(item);
+  };
+
+  return (
+    <Stack
+      style={{ minWidth: '300px' }}
+      flexDirection="column"
+      spacing="spacingM"
+      alignItems="start"
+    >
+      <Autocomplete<Produce>
+        {...args}
+        items={filteredItems}
+        onInputValueChange={handleInputValueChange}
+        onSelectItem={handleSelectItem}
+        itemToString={(item) => item.name}
+        renderItem={(item) => item.name}
+        selectedItem={selectedFruit}
+      />
+
+      <Paragraph>Selected fruit: {selectedFruit?.name}</Paragraph>
+      <Button onClick={() => setSelectedFruit({ id: undefined, name: '' })}>
+        clear selection
+      </Button>
+    </Stack>
+  );
+};
+UsingObjectsAsItems.args = {
+  placeholder: 'Search your favorite fruit',
+};
+
+export const UsingGroupedItems = (args: AutocompleteProps<Produce>) => {
+  const [selectedItem, setSelectedItem] = useState<Produce>({} as Produce);
   const [filteredItems, setFilteredItems] = useState(groceryList);
 
   const handleInputValueChange = (value: string) => {
@@ -173,6 +223,7 @@ export const UsingGroupedItems = (args: AutocompleteProps<GroceryList>) => {
       {/* It’s not necessary to pass "Fruit" (type of one item)  */}
       <Autocomplete<Produce>
         {...args}
+        selectedItem={selectedItem}
         items={filteredItems}
         isGrouped
         renderItem={(item) => item.name}