+++ packages/components/modal/src/Modal.styles.ts
@@ -1,10 +1,11 @@
 import tokens from '@contentful/f36-tokens';
 import { css, cx } from 'emotion';
-import type { ModalPositionType, ModalSizeType } from './types';
+
+import type { ModalProps } from './Modal';
 
 export function getModalStyles(props: {
-  size: ModalSizeType;
-  position: ModalPositionType;
+  size: ModalProps['size'];
+  position: ModalProps['position'];
   allowHeightOverflow?: boolean;
   className?: string;
 }) {
@@ -12,7 +13,7 @@ export function getModalStyles(props: {
     css({
       margin: tokens.spacing2Xl,
       backgroundColor: tokens.colorWhite,
-      borderRadius: tokens.borderRadiusMedium,
+      borderRadius: props.size === 'zen' ? 0 : tokens.borderRadiusMedium,
       boxShadow: tokens.boxShadowHeavy,
       maxHeight: `calc(100vh - 1rem * (100 / ${tokens.fontBaseDefault}))`,
       maxWidth: `calc(100vw - 1rem * (100 / ${tokens.fontBaseDefault}))`,
@@ -52,7 +53,7 @@ export function getModalStyles(props: {
           margin: '0 auto',
           textAlign: 'left',
           outline: 'none',
-          transform: 'scale(0.85)',
+          transform: props.size === 'zen' ? 'scale(1)' : 'scale(0.85)',
           transition: `transform ${tokens.transitionDurationDefault} ${tokens.transitionEasingDefault}`,
         }),
         props.size === 'zen'
@@ -63,10 +64,10 @@ export function getModalStyles(props: {
           : null,
       ),
       afterOpen: css({
-        transform: 'scale(1)',
+        transform: 'scale(1) !important',
       }),
       beforeClose: css({
-        transform: 'scale(0.85)',
+        transform: props.size === 'zen' ? 'scale(1)' : 'scale(0.85)',
       }),
     },
     modalOverlay: {