+++ packages/website/components/LiveEditor/ComponentSource.tsx
@@ -10,7 +10,6 @@ import { css, cx } from 'emotion';
 import { LiveProvider, LiveEditor, LiveError, LivePreview } from 'react-live';
 import tokens from '@contentful/f36-tokens';
 import * as f36Components from '@contentful/f36-components';
-import { Pagination } from '@contentful/f36-pagination';
 import * as f36utils from '@contentful/f36-utils';
 import { useForm, useController } from 'react-hook-form';
 import { MdAccessAlarm } from 'react-icons/md';
@@ -33,7 +32,6 @@ const liveProviderScope = {
   ...f36icons,
   ...f36Components,
   ...f36utils,
-  Pagination,
   css,
   f36icons,
   tokens,