+++ packages/components/autocomplete/CHANGELOG.md
@@ -1,5 +1,137 @@
 # Change Log
 
+## 4.9.0
+
+### Minor Changes
+
+- [#2055](https://github.com/contentful/forma-36/pull/2055) [`910dafc9`](https://github.com/contentful/forma-36/commit/910dafc9ba4a1acc9133d2ee89e7a1fad879e932) Thanks [@Lelith](https://github.com/Lelith)! - Enable Autocomplete as a controlled input by using selectedItem property
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-button@4.9.0
+  - @contentful/f36-forms@4.9.0
+  - @contentful/f36-icons@4.9.0
+  - @contentful/f36-popover@4.9.0
+  - @contentful/f36-skeleton@4.9.0
+  - @contentful/f36-typography@4.9.0
+  - @contentful/f36-utils@4.9.0
+  - @contentful/f36-core@4.9.0
+
+## 4.8.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-button@4.8.2
+  - @contentful/f36-forms@4.8.2
+  - @contentful/f36-icons@4.8.2
+  - @contentful/f36-popover@4.8.2
+  - @contentful/f36-skeleton@4.8.2
+  - @contentful/f36-typography@4.8.2
+  - @contentful/f36-utils@4.8.2
+  - @contentful/f36-core@4.8.2
+
+## 4.8.1
+
+### Patch Changes
+
+- Updated dependencies [[`7afca5e6`](https://github.com/contentful/forma-36/commit/7afca5e6f9d89b1f187fc48ab8e8fd39366f110b)]:
+  - @contentful/f36-typography@4.8.1
+  - @contentful/f36-button@4.8.1
+  - @contentful/f36-forms@4.8.1
+  - @contentful/f36-icons@4.8.1
+  - @contentful/f36-popover@4.8.1
+  - @contentful/f36-skeleton@4.8.1
+  - @contentful/f36-utils@4.8.1
+  - @contentful/f36-core@4.8.1
+
+## 4.8.0
+
+### Patch Changes
+
+- [#2008](https://github.com/contentful/forma-36/pull/2008) [`1a941fd9`](https://github.com/contentful/forma-36/commit/1a941fd99757e49bb30e0e32eab19bc680f66597) Thanks [@massao](https://github.com/massao)! - Fix issue with types being generated wrongly
+
+- Updated dependencies [[`1a941fd9`](https://github.com/contentful/forma-36/commit/1a941fd99757e49bb30e0e32eab19bc680f66597)]:
+  - @contentful/f36-skeleton@4.8.0
+  - @contentful/f36-button@4.8.0
+  - @contentful/f36-forms@4.8.0
+  - @contentful/f36-icons@4.8.0
+  - @contentful/f36-popover@4.8.0
+  - @contentful/f36-typography@4.8.0
+  - @contentful/f36-utils@4.8.0
+  - @contentful/f36-core@4.8.0
+
+## 4.7.0
+
+### Patch Changes
+
+- Updated dependencies [[`e50b8945`](https://github.com/contentful/forma-36/commit/e50b8945faef7323187e1bd9a2a9cbfcbae2e405)]:
+  - @contentful/f36-core@4.7.0
+  - @contentful/f36-button@4.7.0
+  - @contentful/f36-forms@4.7.0
+  - @contentful/f36-icons@4.7.0
+  - @contentful/f36-popover@4.7.0
+  - @contentful/f36-skeleton@4.7.0
+  - @contentful/f36-typography@4.7.0
+  - @contentful/f36-utils@4.7.0
+
+## 4.6.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-button@4.6.2
+  - @contentful/f36-forms@4.6.2
+  - @contentful/f36-icons@4.6.2
+  - @contentful/f36-popover@4.6.2
+  - @contentful/f36-skeleton@4.6.2
+  - @contentful/f36-typography@4.6.2
+  - @contentful/f36-utils@4.6.2
+  - @contentful/f36-core@4.6.2
+
+## 4.6.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-button@4.6.1
+  - @contentful/f36-forms@4.6.1
+  - @contentful/f36-icons@4.6.1
+  - @contentful/f36-popover@4.6.1
+  - @contentful/f36-skeleton@4.6.1
+  - @contentful/f36-typography@4.6.1
+  - @contentful/f36-utils@4.6.1
+  - @contentful/f36-core@4.6.1
+
+## 4.6.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-button@4.6.0
+  - @contentful/f36-forms@4.6.0
+  - @contentful/f36-icons@4.6.0
+  - @contentful/f36-popover@4.6.0
+  - @contentful/f36-skeleton@4.6.0
+  - @contentful/f36-typography@4.6.0
+  - @contentful/f36-utils@4.6.0
+  - @contentful/f36-core@4.6.0
+
+## 4.5.0
+
+### Patch Changes
+
+- Updated dependencies [[`d841d7cb`](https://github.com/contentful/forma-36/commit/d841d7cbf74b20192028591197bc42df3d2ebc2e)]:
+  - @contentful/f36-utils@4.5.0
+  - @contentful/f36-button@4.5.0
+  - @contentful/f36-forms@4.5.0
+  - @contentful/f36-icons@4.5.0
+  - @contentful/f36-popover@4.5.0
+  - @contentful/f36-skeleton@4.5.0
+  - @contentful/f36-typography@4.5.0
+  - @contentful/f36-core@4.5.0
+
 ## 4.4.1
 
 ### Patch Changes