+++ packages/components/modal/CHANGELOG.md
@@ -1,5 +1,105 @@
 # Change Log
 
+## 4.9.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-button@4.9.0
+  - @contentful/f36-icons@4.9.0
+  - @contentful/f36-typography@4.9.0
+  - @contentful/f36-core@4.9.0
+
+## 4.8.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-button@4.8.2
+  - @contentful/f36-icons@4.8.2
+  - @contentful/f36-typography@4.8.2
+  - @contentful/f36-core@4.8.2
+
+## 4.8.1
+
+### Patch Changes
+
+- Updated dependencies [[`7afca5e6`](https://github.com/contentful/forma-36/commit/7afca5e6f9d89b1f187fc48ab8e8fd39366f110b)]:
+  - @contentful/f36-typography@4.8.1
+  - @contentful/f36-button@4.8.1
+  - @contentful/f36-icons@4.8.1
+  - @contentful/f36-core@4.8.1
+
+## 4.8.0
+
+### Patch Changes
+
+- [#2008](https://github.com/contentful/forma-36/pull/2008) [`1a941fd9`](https://github.com/contentful/forma-36/commit/1a941fd99757e49bb30e0e32eab19bc680f66597) Thanks [@massao](https://github.com/massao)! - Fix issue with types being generated wrongly
+
+- Updated dependencies []:
+  - @contentful/f36-button@4.8.0
+  - @contentful/f36-icons@4.8.0
+  - @contentful/f36-typography@4.8.0
+  - @contentful/f36-core@4.8.0
+
+## 4.7.0
+
+### Patch Changes
+
+- Updated dependencies [[`e50b8945`](https://github.com/contentful/forma-36/commit/e50b8945faef7323187e1bd9a2a9cbfcbae2e405)]:
+  - @contentful/f36-core@4.7.0
+  - @contentful/f36-button@4.7.0
+  - @contentful/f36-icons@4.7.0
+  - @contentful/f36-typography@4.7.0
+
+## 4.6.2
+
+### Patch Changes
+
+- [#1994](https://github.com/contentful/forma-36/pull/1994) [`0364ffe2`](https://github.com/contentful/forma-36/commit/0364ffe2aa2e5ccfea6263f4173b2d8069e4ec05) Thanks [@denkristoffer](https://github.com/denkristoffer)! - fix(modal): fix scaling issues for modals with size "zen"
+
+- Updated dependencies []:
+  - @contentful/f36-button@4.6.2
+  - @contentful/f36-icons@4.6.2
+  - @contentful/f36-typography@4.6.2
+  - @contentful/f36-core@4.6.2
+
+## 4.6.1
+
+### Patch Changes
+
+- [#1995](https://github.com/contentful/forma-36/pull/1995) [`f4495eda`](https://github.com/contentful/forma-36/commit/f4495edaf02d87e78acbc454730190161f6a305d) Thanks [@anho](https://github.com/anho)! - force scale for modal
+
+- Updated dependencies []:
+  - @contentful/f36-button@4.6.1
+  - @contentful/f36-icons@4.6.1
+  - @contentful/f36-typography@4.6.1
+  - @contentful/f36-core@4.6.1
+
+## 4.6.0
+
+### Minor Changes
+
+- [`6fd4099e`](https://github.com/contentful/forma-36/commit/6fd4099e3c4f364f6b95cb3720cca995c355e5cc) Thanks [@denkristoffer](https://github.com/denkristoffer)! - fix(modal): custom classname before modal root styles
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-button@4.6.0
+  - @contentful/f36-icons@4.6.0
+  - @contentful/f36-typography@4.6.0
+  - @contentful/f36-core@4.6.0
+
+## 4.5.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-button@4.5.0
+  - @contentful/f36-icons@4.5.0
+  - @contentful/f36-typography@4.5.0
+  - @contentful/f36-core@4.5.0
+
 ## 4.4.1
 
 ### Patch Changes